+++ src/components/single-input-form/types.ts
@@ -1,6 +1,5 @@
 export type SingleInputFormProps = {
   buttonText: string;
   handleClick: (text: string) => Promise<void>;
-  initialValue?: string;
   placeholder?: string;
 };